backend/requirements.txt
pymupdf>=1.24.0
openai>=1.0.0
httpx>=0.27.0
pyyaml>=6.0
rich>=13.0
